refactor(check): extract declarations.ts from checker.ts
Extract type declaration processing functions into a dedicated module:
- getTypeDeclFromLine - Detect type declarations in line nodes
- startTypeDecl - Initialize type declaration context
- processFieldDecl - Process field declarations within types
- finalizeTypeDecl - Register completed type declarations

Also exports internal helpers needed by checker.ts for FieldInit handling:
- getFieldDeclFromLine
- resolveUserDefinedFieldType
- addFieldToTypeDeclContext
