+++ packages/compiler/src/check/checker.ts
@@ -12,8 +12,17 @@
 import type { CompilationContext, StringId } from '../core/context.ts'
 import type { DiagnosticCode } from '../core/diagnostics.ts'
 import { type NodeId, NodeKind, nodeId, offsetNodeId, prevNodeId } from '../core/nodes.ts'
-import { nextTokenId, type Token, TokenKind } from '../core/tokens.ts'
+import { TokenKind } from '../core/tokens.ts'
 import { processVariableBinding } from './bindings.ts'
+import {
+	addFieldToTypeDeclContext,
+	finalizeTypeDecl,
+	getFieldDeclFromLine,
+	getTypeDeclFromLine,
+	processFieldDecl,
+	resolveUserDefinedFieldType,
+	startTypeDecl,
+} from './declarations.ts'
 import { checkExpression } from './expressions.ts'
 import {
 	type BlockContext,
@@ -28,7 +37,6 @@ import {
 	popBlockContext,
 	pushBlockContext,
 	type RecordLiteralContext,
-	type TypeDeclContext,
 } from './state.ts'
 import { InstStore, ScopeStore, SymbolStore, TypeStore } from './stores.ts'
 import { getTypeNameFromToken } from './type-resolution.ts'
@@ -108,165 +116,6 @@ function handleVariableBinding(
 }
 
 /**
- * Get TypeDecl node from a line (if present).
- */
-function getTypeDeclFromLine(
-	lineId: NodeId,
-	context: CompilationContext
-): { id: NodeId; kind: NodeKind } | null {
-	for (const [childId, child] of context.nodes.iterateChildren(lineId)) {
-		if (child.kind === NodeKind.TypeDecl) {
-			return { id: childId, kind: child.kind }
-		}
-	}
-	return null
-}
-
-/**
- * Get FieldDecl node from a line (if present).
- */
-function getFieldDeclFromLine(
-	lineId: NodeId,
-	context: CompilationContext
-): { id: NodeId; kind: NodeKind } | null {
-	for (const [childId, child] of context.nodes.iterateChildren(lineId)) {
-		if (child.kind === NodeKind.FieldDecl) {
-			return { id: childId, kind: child.kind }
-		}
-	}
-	return null
-}
-
-/**
- * Start processing a type declaration.
- * Extracts the type name and initializes the context for collecting fields.
- */
-function startTypeDecl(typeDeclId: NodeId, state: CheckerState, context: CompilationContext): void {
-	const typeDeclNode = context.nodes.get(typeDeclId)
-	// TypeDecl node's tokenId points to the 'type' keyword
-	// The type name is in the next token (the identifier)
-	const identTokenId = nextTokenId(typeDeclNode.tokenId)
-	const identToken = context.tokens.get(identTokenId)
-	const typeName = context.strings.get(identToken.payload as StringId)
-
-	pushBlockContext(state, {
-		children: [],
-		expectedChildKind: NodeKind.FieldDecl,
-		fieldNames: new Set(),
-		fields: [],
-		kind: 'TypeDecl',
-		nodeId: typeDeclId,
-		typeId: null, // Will be assigned after registration
-		typeName,
-	})
-}
-
-/**
- * Resolves a user-defined field type by name lookup.
- */
-function resolveUserDefinedFieldType(
-	fieldTypeName: string,
-	fieldDeclId: NodeId,
-	state: CheckerState,
-	context: CompilationContext
-): TypeId | null {
-	const ctx = currentBlockContext(state)
-	if (ctx?.kind === 'TypeDecl' && fieldTypeName === ctx.typeName) {
-		context.emitAtNode('TWCHECK032' as DiagnosticCode, fieldDeclId, {
-			field: fieldTypeName,
-			type: fieldTypeName,
-		})
-		return null
-	}
-
-	const lookedUpTypeId = state.types.lookup(fieldTypeName)
-	if (lookedUpTypeId === undefined) {
-		context.emitAtNode('TWCHECK010' as DiagnosticCode, fieldDeclId, {
-			name: fieldTypeName,
-		})
-		return null
-	}
-	return lookedUpTypeId
-}
-
-/**
- * Resolves field type from token (either user-defined or primitive).
- * Returns typeId or null if type is invalid.
- */
-function resolveFieldType(
-	typeToken: Token,
-	fieldDeclId: NodeId,
-	state: CheckerState,
-	context: CompilationContext
-): TypeId | null {
-	if (typeToken.kind === TokenKind.Identifier) {
-		const fieldTypeName = context.strings.get(typeToken.payload as StringId)
-		return resolveUserDefinedFieldType(fieldTypeName, fieldDeclId, state, context)
-	}
-
-	const typeInfo = getTypeNameFromToken(typeToken.kind)
-	if (!typeInfo) {
-		context.emitAtNode('TWCHECK010' as DiagnosticCode, fieldDeclId, {
-			name: 'unknown',
-		})
-		return null
-	}
-	return typeInfo.typeId
-}
-
-/**
- * Process a FieldDecl node within a type declaration.
- * Extracts field name and type, checking for duplicates.
- */
-function processFieldDecl(
-	fieldDeclId: NodeId,
-	state: CheckerState,
-	context: CompilationContext
-): void {
-	const ctx = currentBlockContext(state)
-	if (!ctx || ctx.kind !== 'TypeDecl') {
-		// FieldDecl outside type declaration - should not happen if parser is correct
-		return
-	}
-
-	const fieldDeclNode = context.nodes.get(fieldDeclId)
-	const fieldToken = context.tokens.get(fieldDeclNode.tokenId)
-	const fieldName = context.strings.get(fieldToken.payload as StringId)
-
-	// Get the type token (field token + 2: skip colon)
-	// Token layout: Identifier, Colon, TypeKeyword/Identifier
-	const typeTokenId = (fieldDeclNode.tokenId as number) + 2
-	const typeToken = context.tokens.get(typeTokenId as typeof fieldDeclNode.tokenId)
-
-	const fieldTypeId = resolveFieldType(typeToken, fieldDeclId, state, context)
-	if (!fieldTypeId) return
-
-	addFieldToTypeDeclContext(ctx, fieldName, fieldTypeId, fieldDeclId, context)
-}
-
-/**
- * Finalize a type declaration by registering it with the TypeStore.
- */
-function finalizeTypeDecl(state: CheckerState, _context: CompilationContext): void {
-	const ctx = currentBlockContext(state)
-	if (!ctx || ctx.kind !== 'TypeDecl') return
-
-	popBlockContext(state)
-
-	const { fields, nodeId, typeName } = ctx
-	if (!fields) return
-
-	// Convert to FieldInfo format
-	const fieldInfos = fields.map((f, index) => ({
-		index,
-		name: f.name,
-		typeId: f.typeId,
-	}))
-
-	state.types.registerRecordType(typeName, fieldInfos, nodeId)
-}
-
-/**
  * Get FieldInit node from a line (if present).
  */
 function getFieldInitFromLine(
@@ -335,29 +184,6 @@ function getNestedRecordInitFromFieldInit(
 		}
 	}
 	return null
-}
-
-/**
- * Add a field to the current TypeDecl block context.
- * Returns false if the field name is a duplicate.
- */
-function addFieldToTypeDeclContext(
-	ctx: TypeDeclContext,
-	fieldName: string,
-	fieldTypeId: TypeId,
-	fieldNodeId: NodeId,
-	context: CompilationContext
-): boolean {
-	if (ctx.fieldNames.has(fieldName)) {
-		context.emitAtNode('TWCHECK026' as DiagnosticCode, fieldNodeId, {
-			name: fieldName,
-			typeName: ctx.typeName,
-		})
-		return false
-	}
-	ctx.fieldNames.add(fieldName)
-	ctx.fields.push({ name: fieldName, nodeId: fieldNodeId, typeId: fieldTypeId })
-	return true
 }
 
 /**
